refs #151 Add i18next and translate main menu
