+++ package.json
@@ -104,6 +104,9 @@
     "element-ui": "^2.3.4",
     "hawk": "^7.0.7",
     "hoek": "^5.0.3",
+    "i18next": "^11.5.0",
+    "i18next-node-fs-backend": "^1.2.1",
+    "i18next-sync-fs-backend": "^1.1.0",
     "is-empty": "^1.2.0",
     "lodash": "^4.17.10",
     "megalodon": "^0.3.0",
