+++ src/main/index.js
@@ -7,7 +7,6 @@ import log from 'electron-log'
 import windowStateKeeper from 'electron-window-state'
 import simplayer from 'simplayer'
 import path from 'path'
-import openAboutWindow from 'about-window'
 import ContextMenu from 'electron-context-menu'
 import * as Splashscreen from '@trodi/electron-splashscreen'
 
@@ -16,6 +15,8 @@ import Account from './account'
 import Streaming from './streaming'
 import Preferences from './preferences'
 import Hashtags from './hashtags'
+import ApplicationMenu from './menu'
+import i18n from '../config/i18n'
 
 /**
  * Context menu
@@ -107,165 +108,12 @@ async function createWindow () {
       click: () => changeAccount(a, index)
     }
   })
-  /**
-   * For mac menu
-   */
-  const macGeneralMenu = process.platform !== 'darwin' ? [] : [
-    {
-      type: 'separator'
-    },
-    {
-      label: 'Services',
-      role: 'services',
-      submenu: []
-    },
-    {
-      type: 'separator'
-    },
-    {
-      label: 'Hide Whalebird',
-      role: 'hide'
-    },
-    {
-      label: 'Hide Others',
-      role: 'hideothers'
-    },
-    {
-      label: 'Show All',
-      role: 'unhide'
-    }
-  ]
+
+  i18n.changeLanguage('ja')
   /**
    * Set application menu
    */
-  const template = [
-    {
-      label: 'Whalebird',
-      submenu: [
-        {
-          label: 'About Whalebird',
-          role: 'about',
-          click: () => {
-            openAboutWindow({
-              icon_path: path.resolve(__dirname, '../../build/icons/256x256.png'),
-              copyright: 'Copyright (c) 2018 AkiraFukushima',
-              package_json_dir: path.resolve(__dirname, '../../'),
-              open_devtools: process.env.NODE_ENV !== 'production'
-            })
-          }
-        },
-        {
-          type: 'separator'
-        },
-        {
-          label: 'Preferences...',
-          accelerator: 'CmdOrCtrl+,',
-          click: () => {
-            mainWindow.webContents.send('open-preferences')
-          }
-        },
-        ...macGeneralMenu,
-        {
-          type: 'separator'
-        },
-        {
-          label: 'Quit',
-          accelerator: 'CmdOrCtrl+Q',
-          role: 'quit'
-        }
-      ]
-    },
-    {
-      label: 'Toot',
-      submenu: [
-        {
-          label: 'New Toot',
-          accelerator: 'CmdOrCtrl+N',
-          click: () => {
-            mainWindow.webContents.send('CmdOrCtrl+N')
-          }
-        }
-      ]
-    },
-    {
-      label: 'Edit',
-      submenu: [
-        {
-          label: 'Undo',
-          accelerator: 'CmdOrCtrl+Z',
-          role: 'undo'
-        },
-        {
-          label: 'Redo',
-          accelerator: 'Shift+CmdOrCtrl+Z',
-          role: 'redo'
-        },
-        {
-          type: 'separator'
-        },
-        {
-          label: 'Cut',
-          accelerator: 'CmdOrCtrl+X',
-          role: 'cut'
-        },
-        {
-          label: 'Copy',
-          accelerator: 'CmdOrCtrl+C',
-          role: 'copy'
-        },
-        {
-          label: 'Paste',
-          accelerator: 'CmdOrCtrl+V',
-          role: 'paste'
-        },
-        {
-          label: 'Select All',
-          accelerator: 'CmdOrCtrl+A',
-          role: 'selectall'
-        }
-      ]
-    },
-    {
-      label: 'View',
-      submenu: [
-        {
-          label: 'Toggle Full Screen',
-          role: 'togglefullscreen'
-        }
-      ]
-    },
-    {
-      label: 'Window',
-      submenu: [
-        {
-          label: 'Close Window',
-          role: 'close'
-        },
-        {
-          label: 'Minimize',
-          role: 'minimize'
-        },
-        {
-          type: 'separator'
-        },
-        {
-          label: 'Jump to',
-          accelerator: 'CmdOrCtrl+K',
-          enabled: true,
-          click: () => {
-            mainWindow.webContents.send('CmdOrCtrl+K')
-          }
-        },
-        {
-          type: 'separator'
-        },
-        ...accountsChange
-      ]
-    }
-  ]
-
-  const menu = Menu.buildFromTemplate(template)
-  Menu.setApplicationMenu(menu)
+  ApplicationMenu(mainWindow, accountsChange, i18n)
 
   /**
    * Set dock menu for mac
